% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloadr.R
\name{myloadr}
\alias{myloadr}
\title{Attach libraries}
\usage{
myloadr(..., update = FALSE)
}
\arguments{
\item{...}{...}

\item{update}{...}
}
\value{
...
}
\description{
Attach libraries
}
\examples{
loadr(stats)
}
